Create shellcoderunnerTwo.vba
This is the second payload that I used in cow motors and have tested to great effect. 

This solves a big issue i noticed in that excessive use of encryption as well as process injection seems to get caught out more often than using sleep. I also have the msfvenom command migrate to explorer before i get a shell back